v2 Operaciones con coma flotante añadida

+++ src/Clases/Calculadora.java
@@ -1,19 +1,19 @@
 package Clases;
 
 public class Calculadora {
-	public int suma(int num1, int num2) {
+	public float suma(float num1, float num2) {
 		return num1+num2;
 	}
 	
-	public int resta(int num1, int num2) {
+	public float resta(float num1, float num2) {
 		return num1-num2;
 	}
 	
-	public int multiplicacion(int num1, int num2) {
+	public float multiplicacion(float num1, float num2) {
 		return num1*num2;
 	}
 	
-	public int division(int num1, int num2) {
+	public float division(float num1, float num2) {
 		return num1/num2;
 	}
 }
